// frontend/src/redux/features/types.d.ts
export interface ConversationsType {
  _id: string;
  id: number;
  status: string;
  client: boolean;
  time?: any;
  selected: false;
  deadline?: any;
  hours?: number;
  minutes?: number;
  seconds?: number;
  timer?: any;
  messageDateTime: string;
  online: boolean;
  name: string;
  messages: MessageType[];
}
export type ChatMessageType = {
  text: string;
  is_bot: boolean;
  is_admin: boolean;
  date?: string;
  is_read?: boolean;
  chat_status: string;
  scenario_name: string;
};

export type ChatConversationType = {
  // _id?: string;
  phone_number: string;
  chat_messages: ChatMessageType[];
  unread_msg?: number;
};

export type ChatsByCompanyReturnType = {
  data?: any;
  status?: number;

  _id: string;
  phone_number_id: string;
  company?: string;
  conversations: ChatConversationType[];
};

export interface MessageType {
  id: number;
  message: string;
  messageTime: string;
  client: boolean;
  name: string;
}
// ...
export type TypeNewConversation={
    _id: string;
    phone_number: string;
    status: string;
    chat_ids: string[];
    credential_id: string;
    createdAt: string;
    updatedAt: string;
    __v: 0;
  }
export interface ChatBotType {
  sideBarToggle: boolean;
  displayChatUiToggle: boolean;
  displayClientInfoInChatToggle: boolean;
  selectedConversation: number;
  selectedConversationObj: ConversationsType;
  conversations: ConversationsType[];
  chats: MessageType[];
  clearTimerInterval?: boolean;
  firstTime?: boolean;
  companychats: ChatsByCompanyReturnType;
  chatsConversation: ChatConversationType[];
  message: ChatConversationType;
  filterStatus: string;
  conversationStatus: string | any;
  currentChatStatus: string;
  arrayStatus: {
    id: string;
    status: string;
    color: string;
  }[];
  selectedStatus: string;
  companyId: string;
  loadingStatus: boolean;
  conversation: TypeNewConversation[];
  chatConversation: {
    _id: string;
    is_read: boolean;
    text: string;
    origin: string;
    converastion_id: string;
    url: string;
    createdAt: string;
    updatedAt: string;
    __v: 0;
  }[];
  allconversation: {
    id: string;
    chats: {
      _id: string;
      is_read: boolean;
      text: string;
      origin: string;
      converastion_id: string;
      url: string;
      createdAt: string;
      updatedAt: string;
      __v: 0;
    }[];
  }[];
}